refactored to allow each event listener to live within its corresponding module; fixed issue of wrong classes being used in some cases due to copy/paste from one component module to the next.
