+++ scripts/news/NewsList.js
@@ -13,3 +13,19 @@ const render = newsCollection => {
         </article>
     `
 }
+/*
+*   Listens for a custom event, backgroundColorChosen, from the eventHub (#container),
+*   removes any class list associated with color themes, and applies the value sent as a class.
+*/
+export const changeNewsBackgroundColor = () => {
+    const eventHub = document.querySelector("#container");
+    const targetFavoriteElementBackgroundColorArray = document.querySelectorAll(".newsItem");
+
+    eventHub.addEventListener('backgroundColorChosen', event => {
+        targetFavoriteElementBackgroundColorArray.forEach(e => {
+                e.classList.remove('blue', 'red', 'green', 'purple');
+                e.classList.add(event.detail.color);
+            })
+        }
+    )
+}
